BF4 experiments, caching actually works now

+++ Frostbite/IO.cs
@@ -23,43 +23,46 @@ public class IO
     /// </summary>
     public static void DecryptAndCache(string path)
     {
-        using (var r = new BinaryReader(File.OpenRead(path)))
+        if (!Path.Exists(path))
         {
-            byte[] magic = r.ReadBytes(4);
-            byte[] data;
-
-            // Is XOR encrypted.
-            if (magic.SequenceEqual(new byte[] { 0x00, 0xD1, 0xCE, 0x00 }) || 
-                magic.SequenceEqual(new byte[] { 0x00, 0xD1, 0xCE, 0x01 }))
+            using (var r = new BinaryReader(File.OpenRead(path)))
             {
-                r.BaseStream.Position = 296; // Skip the signature.
-                var key = r.ReadBytes(260);
-                for (int i = 0; i < key.Length; i++)
+                byte[] magic = r.ReadBytes(4);
+                byte[] data;
+
+                // Is XOR encrypted.
+                if (magic.SequenceEqual(new byte[] { 0x00, 0xD1, 0xCE, 0x00 }) ||
+                    magic.SequenceEqual(new byte[] { 0x00, 0xD1, 0xCE, 0x01 }))
                 {
-                    key[i] ^= 0x7B; // XOR with 0x7B (Bytes 257, 258 and 259 are unused).
+                    r.BaseStream.Position = 296; // Skip the signature.
+                    var key = r.ReadBytes(260);
+                    for (int i = 0; i < key.Length; i++)
+                    {
+                        key[i] ^= 0x7B; // XOR with 0x7B (Bytes 257, 258 and 259 are unused).
+                    }
+                    byte[] encryptedData = r.ReadUntilStreamEnd();
+                    data = new byte[encryptedData.Length];
+                    for (int i = 0; i < encryptedData.Length; i++)
+                        data[i] = (byte)(key[i % 257] ^ encryptedData[i]);
                 }
-                byte[] encryptedData = r.ReadUntilStreamEnd();
-                data = new byte[encryptedData.Length];
-                for (int i = 0; i < encryptedData.Length; i++)
-                    data[i] = (byte)(key[i % 257] ^ encryptedData[i]);
-            }
-            // Is not XOR encrypted but has sequence + key.
-            else if (magic.SequenceEqual(new byte[] { 0x00, 0xD1, 0xCE, 0x03 }))
-            {
-                r.BaseStream.Position = 296; // Skip the signature.
-                r.ReadBytes(260); // Empty key.
-                data = r.ReadUntilStreamEnd();
-            }
-            // Not encrypted.
-            else
-            {
-                r.BaseStream.Position = 0; // Go back to the start of the file;
-                data = r.ReadUntilStreamEnd(); // Read data.
-            }
+                // Is not XOR encrypted but has sequence + key.
+                else if (magic.SequenceEqual(new byte[] { 0x00, 0xD1, 0xCE, 0x03 }))
+                {
+                    r.BaseStream.Position = 296; // Skip the signature.
+                    r.ReadBytes(260); // Empty key.
+                    data = r.ReadUntilStreamEnd();
+                }
+                // Not encrypted.
+                else
+                {
+                    r.BaseStream.Position = 0; // Go back to the start of the file;
+                    data = r.ReadUntilStreamEnd(); // Read data.
+                }
 
-            // Write the Catalog to file to cache it.
-            Directory.CreateDirectory($"Cache\\{Settings.CurrentGame}");
-            File.WriteAllBytes($"Cache\\{Settings.CurrentGame}\\{Path.GetFileName(path)}", data);
+                // Write the Catalog to file to cache it.
+                Directory.CreateDirectory($"Cache\\{Settings.CurrentGame}");
+                File.WriteAllBytes($"Cache\\{Settings.CurrentGame}\\{Path.GetFileName(path)}", data);
+            }
         }
     }
 }
